fix: 투표 주제 조회 제외
